% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{check_cmethod}
\alias{check_cmethod}
\title{Check correlation method}
\usage{
check_cmethod(c_method)
}
\arguments{
\item{c_method}{Normalization method}
}
\value{
true or false
}
\description{
Check correlation method
}

