% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{check_expression_list}
\alias{check_expression_list}
\title{Check expression gene list}
\usage{
check_expression_list(exp_list)
}
\arguments{
\item{exp_list}{Expression gene list}
}
\value{
true or false
}
\description{
Check expression gene list
}

