% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{check_hkgene_list}
\alias{check_hkgene_list}
\title{Check housekeeping gene list}
\usage{
check_hkgene_list(hkgene_list)
}
\arguments{
\item{hkgene_list}{Housekeeping gene list}
}
\value{
true or false
}
\description{
Check housekeeping gene list
}

