% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{check_method}
\alias{check_method}
\title{Check normalization method}
\usage{
check_method(norm_method)
}
\arguments{
\item{norm_method}{Normalization method}
}
\value{
true or false
}
\description{
Check normalization method
}

