% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{check_norm_list}
\alias{check_norm_list}
\title{Check normalized expression gene list}
\usage{
check_norm_list(norm_list)
}
\arguments{
\item{norm_list}{Normalized expression gene list}
}
\value{
true or false
}
\description{
Check normalized expression gene list
}

