% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{exists_hkgene_list}
\alias{exists_hkgene_list}
\title{Check if housekeeping gene list is valid}
\usage{
exists_hkgene_list(NVTobject, hkgene_list)
}
\arguments{
\item{NVTobject}{A initialized NVTdata object}

\item{hkgene_list}{Housekeeping gene list}
}
\value{
true or false
}
\description{
Check if housekeeping gene list is valid
}

