% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\docType{data}
\name{myexp1}
\alias{myexp1}
\title{List of expression data of each gene}
\format{A data frame with 896 rows and 1 variable:
\describe{
  \item{Mapped_reads_Hypoxy}{gene expression in mapped reads}
  ...
}}
\usage{
myexp1
}
\description{
A data set with each gene name and its expression in mapped reads, under hypoxic condictions
}
\keyword{datasets}

