% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\docType{data}
\name{mylen}
\alias{mylen}
\title{List of lenght of each gene}
\format{A data frame with 896 rows and 1 variable:
\describe{
  \item{length}{gene lenght in bp}
  ...
}}
\usage{
mylen
}
\description{
A data set with each gene name and its length in nucleotides
}
\keyword{datasets}

