% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\docType{data}
\name{mylist_hs}
\alias{mylist_hs}
\title{List of human housekeeping genes}
\format{A list of 10 elements:
\itemize{
  \item{HK_genes}
}}
\usage{
mylist_hs
}
\description{
A data set of 10 human housekeeping genes
}
\keyword{datasets}

