% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\docType{data}
\name{myusecaseexp}
\alias{myusecaseexp}
\title{Dataframe of expression of six human RNA-Seq samples}
\format{A data frame with 60619 rows and 6 variables:
\itemize{
  \item{GSM1464282}
  \item{GSM1464283}
  \item{GSM1464284}
  \item{GSM1464289}
  \item{GSM1464290}
  \item{GSM1464291}
}}
\usage{
myusecaseexp
}
\description{
A data set with each gene name and its expression in the six samples
}
\keyword{datasets}

