% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\docType{data}
\name{myusecaselen}
\alias{myusecaselen}
\title{List of ensembl human gene-length}
\format{A list of 60619 elements:
\itemize{
  \item{Length}
}}
\usage{
myusecaselen
}
\description{
A data set with each human ensemble gene name and its length in nucleotides
}
\keyword{datasets}

