% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-classes.R
\docType{class}
\name{NVTdata class}
\alias{NVT}
\alias{NVT-class}
\alias{NVTdata class}
\title{Class "NVTdata"}
\description{
A class with all required NVTdata
}
\section{Objects from the Class}{
 Objects can be created by calls of the
form \code{new("NVT", ...)}.
}
\examples{

showClass("NVTdata")

}
\author{
Thomas Eder
}
\keyword{classes}

