% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTlm}
\alias{NVTlm}
\title{Returns the linear model a normalized NVTobject}
\usage{
NVTlm(NVTdataobj)
}
\arguments{
\item{NVTdataobj}{A previously initialized and normalized NVTobject}
}
\value{
Returns the linear model a normalized NVTobject
}
\description{
Returns the linear model a normalized NVTobject
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1<-c("ENSG00000111640","ENSG00000163631","ENSG00000075624","ENSG00000172053",
"ENSG00000170950","ENSG00000165704","ENSG00000196839","ENSG00000168938","ENSG00000177700")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")
mynorm <- NVTnormalize(mynvt)

mylm <- NVTlm(mynorm)

summary(mylm)
}

