% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTloadgff}
\alias{NVTloadgff}
\title{Load gff2 (gtf) or gff3 file}
\usage{
NVTloadgff(gff_file, gff_version, gff_feature, gff_name)
}
\arguments{
\item{gff_file}{The annotation in gff format}

\item{gff_version}{The version of the provided gff file [gff1,gff2,gff3,gtf]}

\item{gff_feature}{The feature to use [default: exon]}

\item{gff_name}{The name to use [default: gene_id]}
}
\value{
List of gene/exon names and their length
}
\description{
Load gff2 (gtf) or gff3 file
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
#get test GFF-file provided by this library
mygffpath<-system.file("extdata", "Ctr-D-UW3CX.gff", package = "NVT")

mylen <- NVTloadgff(mygffpath,"gff3","gene","locus_tag")
mylist1 <- c("CT462","CT115","CT045","CT678")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N",mylen)
}

