% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTmae}
\alias{NVTmae}
\title{Calculate the mean absolute error (MAE) of the housekeeping genes of an initialized and normalized NVTobject}
\usage{
NVTmae(NVTdataobj)
}
\arguments{
\item{NVTdataobj}{A previously initialized and normalized NVTobject}
}
\value{
mean absolute error (MAE) of the normalized housekeeping genes between the samples
}
\description{
Calculate the mean absolute error (MAE) of the housekeeping genes of an initialized and normalized NVTobject
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1=c("CT462","CT115","CT045","CT678")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"TMM")
mynorm <- NVTnormalize(mynvt)

NVTmae(mynorm)
}

