% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTpearson}
\alias{NVTpearson}
\title{Calculate the pearson correclation of the housekeeping genes of an initialized and normalized NVTobject}
\usage{
NVTpearson(NVTdataobj)
}
\arguments{
\item{NVTdataobj}{A previously initialized and normalized NVTobject}
}
\value{
Pearson correlation of the normalized housekeeping genes between the samples
}
\description{
Calculate the pearson correclation of the housekeeping genes of an initialized and normalized NVTobject
}
\examples{
myexp1 = read.table("exp1.txt")
myexp2 = read.table("exp2.txt")
mylen = read.table("length.txt")
mylist1=c('gene1','gene2','gene3','gene10')

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")
mynorm <- NVTnormalize(mynvt)

NVTpearson(mynorm)
}

