% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTtestall}
\alias{NVTtestall}
\title{Calculate the chosen correclation of the housekeeping genes of an initialized NVTobject for all normalization methods}
\usage{
NVTtestall(NVTdataobj, cmethod)
}
\arguments{
\item{NVTdataobj}{A previously initialzed and normalized NVTobject}

\item{cmethod}{Pearson correlation (p), root mean square deviation (rmsd) or mean absolute error (mae)}
}
\value{
Sorted pearson correlation coefficients, RMSD or MAE of the normalized housekeeping genes between the samples
}
\description{
Calculate the chosen correclation of the housekeeping genes of an initialized NVTobject for all normalization methods
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1<-c("ENSG00000111640","ENSG00000163631","ENSG00000075624","ENSG00000172053",
"ENSG00000165704","ENSG00000196839","ENSG00000168938","ENSG00000177700")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N",mylen)

NVTtestall(mynvt,"p")
}

