% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\docType{data}
\name{myexp1}
\alias{myexp1}
\title{List of expression data of each gene}
\format{A data frame with 64102 rows and 1 variable:
\itemize{
  \item{GSM1275862}
}}
\usage{
myexp1
}
\description{
A data set with each gene name and its expression
}
\keyword{datasets}

