% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\docType{data}
\name{mylen}
\alias{mylen}
\title{List of length of each gene}
\format{A data frame with 64102 rows and 1 variable:
\itemize{
  \item{Length}
}}
\usage{
mylen
}
\description{
A data set with each gene name and its length in nucleotides
}
\keyword{datasets}

