% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTadvancedplot}
\alias{NVTadvancedplot}
\title{Plot the MA-plot of a NVTobject with ggplot2}
\usage{
NVTadvancedplot(NVTdataobj)
}
\arguments{
\item{NVTdataobj}{A previously initialized and normalized NVTobject}
}
\value{
Plots the MA-plot with the housekeeping genes indicated
}
\description{
Plot the MA-plot of a NVTobject with ggplot2
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1=c("CT462","CT115","CT045","CT678")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")
mynorm <- NVTnormalize(mynvt)

NVTadvancedplot(mynorm)
}

