% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTadvancedplot}
\alias{NVTadvancedplot}
\title{Plot the XY-plot of a NVTobject with ggplot2}
\usage{
NVTadvancedplot(NVTdataobj, p_cex = 1, t_cex = 1, l_cex = 1)
}
\arguments{
\item{NVTdataobj}{A previously initialized and normalized NVTobject}

\item{p_cex}{Point size factor [default = 1]}

\item{t_cex}{Title size factor [default = 1]}

\item{l_cex}{Label size factor [default = 1]}
}
\value{
Plots the Scatter-plot with the housekeeping genes indicated
}
\description{
Plot the XY-plot of a NVTobject with ggplot2
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1<-c("ENSG00000111640","ENSG00000163631","ENSG00000075624","ENSG00000172053",
"ENSG00000165704","ENSG00000196839","ENSG00000168938","ENSG00000177700")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")
mynorm <- NVTnormalize(mynvt)

NVTadvancedplot(mynorm,2,2,2)
}

