% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTinit}
\alias{NVTinit}
\title{initialize and load input dato into NVTobject}
\usage{
NVTinit(hkgene_list, exp_list1, exp_list2, method, length)
}
\arguments{
\item{hkgene_list}{A list of housekeeping-genes}

\item{exp_list1}{The first data frame of expression values per gene}

\item{exp_list2}{Second data frame of expression values per gene}

\item{method}{The normalization method to us [N,TC,Med,TMM,UQ,UQ2,Q,RPKM,RPM,DEQ,TPM,G]}

\item{length}{A data frame of length per gene}
}
\value{
A NVTobject ready for normalization
}
\description{
initialize and load input dato into NVTobject
}
\examples{
myexp1 = read.table("exp1.txt")
myexp2 = read.table("exp2.txt")
mylen = read.table("length.txt")
mylist1=c('gene1','gene2','gene3','gene10')

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")
mynvt2 <- NVTinit(mylist1,myexp1,myexp2,"RPKM",mylen)
}

