% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTinit}
\alias{NVTinit}
\title{Initialize and load input data into a NVTobject}
\usage{
NVTinit(hkgene_list, exp_list1, exp_list2, nmethod, len)
}
\arguments{
\item{hkgene_list}{A list of housekeeping-genes}

\item{exp_list1}{The first data frame of expression values per gene}

\item{exp_list2}{Second data frame of expression values per gene}

\item{nmethod}{The normalization method to use [N,TC,Med,TMM,UQ,UQ2,Q,RPKM,RPM,DEQ,TPM,G]}

\item{len}{A data frame of length per gene}
}
\value{
A NVTobject ready for normalization
}
\description{
Initialize and load input data into a NVTobject
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1=c("CT462","CT115","CT045","CT678")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")
mynvt2 <- NVTinit(mylist1,myexp1,myexp2,"RPKM",mylen)
}

