% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTnormalize}
\alias{NVTnormalize}
\title{Normalize a NVTobject, the normalization method has been set already in the initialization step}
\usage{
NVTnormalize(NVTdataobj, verbose = TRUE)
}
\arguments{
\item{NVTdataobj}{A previously initialized NVTobject}

\item{verbose}{mode on or off (T/F) [default=TRUE]}
}
\value{
A normalized NVTobject
}
\description{
Normalize a NVTobject, the normalization method has been set already in the initialization step
}
\examples{
library("NVT")

data(myexp1)
data(myexp2)
data(mylen)
mylist1<-c("ENSG00000111640","ENSG00000163631","ENSG00000075624","ENSG00000172053",
"ENSG00000170950","ENSG00000165704","ENSG00000196839","ENSG00000168938","ENSG00000177700")


mynvt <- NVTinit(mylist1,myexp1,myexp2,"N",mylen)

mynorm <- NVTnormalize(mynvt)
}

