% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTnormalize}
\alias{NVTnormalize}
\title{Normalize a NVTobject,
the normalization method has been set already in the initialization step}
\usage{
NVTnormalize(NVTdataobj)
}
\arguments{
\item{NVTdataobj}{A previously initialized NVTobject}
}
\value{
A normalized NVTobject
}
\description{
Normalize a NVTobject,
the normalization method has been set already in the initialization step
}
\examples{
myexp1 = read.table("data/exp1.txt")
myexp2 = read.table("data/exp2.txt")
mylen = read.table("data/length.txt")
mylist1=c('gene1','gene2','gene3','gene10')

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")

mynorm <- NVTnormalize(mynvt)
}

