% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTnormalize}
\alias{NVTnormalize}
\title{Normalize a NVTobject,
the normalization method has been set already in the initialization step}
\usage{
NVTnormalize(NVTdataobj)
}
\arguments{
\item{NVTdataobj}{A previously initialized NVTobject}
}
\value{
A normalized NVTobject
}
\description{
Normalize a NVTobject,
the normalization method has been set already in the initialization step
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1=c("CT462","CT115","CT045","CT678")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N",mylen)

mynorm <- NVTnormalize(mynvt)
}

