% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTplot}
\alias{NVTplot}
\title{Plot the MA-plot of a NVTobject}
\usage{
NVTplot(NVTdataobj)
}
\arguments{
\item{NVTdataobj}{A previously initialized and normalized NVTobject}
}
\value{
Plots the MA-plot with the housekeeping genes indicated
}
\description{
Plot the MA-plot of a NVTobject
}
\examples{
myexp1 = read.table("exp1.txt")
myexp2 = read.table("exp2.txt")
mylen = read.table("length.txt")
mylist1=c('gene1','gene2','gene3','gene10')

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")
mynorm <- NVTnormalize(mynvt)

NVTplot(mynorm)
}

