% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTplot}
\alias{NVTplot}
\title{Plot the MA-plot of a NVTobject}
\usage{
NVTplot(NVTdataobj, cex = 1, ...)
}
\arguments{
\item{NVTdataobj}{A previously initialized and normalized NVTobject}

\item{cex}{Scaling of points and text relative to the default [default=1]}

\item{...}{Arguments passed on to the plot function}
}
\value{
Plots the MA-plot with the housekeeping genes indicated
}
\description{
Plot the MA-plot of a NVTobject
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1<-c("ENSG00000111640","ENSG00000163631","ENSG00000075624","ENSG00000172053",
"ENSG00000170950","ENSG00000165704","ENSG00000196839","ENSG00000168938","ENSG00000177700")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")
mynorm <- NVTnormalize(mynvt)

NVTplot(mynorm,0.8)
}

