% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTrmsd}
\alias{NVTrmsd}
\title{Calculate the root mean squared deviation (RMSD) of the housekeeping genes of an initialized and normalized NVTobject}
\usage{
NVTrmsd(NVTdataobj)
}
\arguments{
\item{NVTdataobj}{A previously initialized and normalized NVTobject}
}
\value{
Root mean squared deviation (RMSD) of the normalized housekeeping genes between the samples
}
\description{
Calculate the root mean squared deviation (RMSD) of the housekeeping genes of an initialized and normalized NVTobject
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1<-c("ENSG00000111640","ENSG00000163631","ENSG00000075624","ENSG00000172053",
"ENSG00000170950","ENSG00000165704","ENSG00000196839","ENSG00000168938","ENSG00000177700")

mynvt <- NVTinit(mylist1,myexp1,myexp2,"TMM")
mynorm <- NVTnormalize(mynvt)

NVTrmsd(mynorm)
}

