% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVT-init.R
\name{NVTtestall}
\alias{NVTtestall}
\title{Calculate the pearson correclation of the housekeeping genes of an initialized NVTobject for all normalization methods}
\usage{
NVTtestall(NVTdataobj)
}
\arguments{
\item{NVTdataobj}{A previously initialzed and normalized NVTobject}
}
\value{
Sorted pearson correlations of the normalized housekeeping genes between the samples
}
\description{
Calculate the pearson correclation of the housekeeping genes of an initialized NVTobject for all normalization methods
}
\examples{
library("NVT")
data(myexp1)
data(myexp2)
data(mylen)
mylist1=c('gene1','gene2','gene3','gene10')

mynvt <- NVTinit(mylist1,myexp1,myexp2,"N")

NVTtestall(mynvt)
}

